requirements.txt
pandas
matplotlib
numpy
hydra-core
wandb
pytorch-lightning
capymoa
timm